# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Powered Smart Email Assistant",
    "description": "AI-Powered Smart Email Assistant REST API with Gemini Generative AI, OCR Document Scanner, and NLP Classification.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/emails/analyze": {
      "post": {
        "tags": [
          "Emails"
        ],
        "summary": "Analyze Pasted Email",
        "description": "Analyse an email the user pasted, then optionally store it.\n\nThe message does not have to come from a connected mailbox, which is the\npoint: forwarded threads, text copied out of a screenshot and mail from a\nprovider you have not linked all work.",
        "operationId": "analyze_pasted_email_api_emails_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/emails/history": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "Get History",
        "description": "Everything this account has analysed, newest first.",
        "operationId": "get_history_api_emails_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EmailItem"
                  },
                  "type": "array",
                  "title": "Response Get History Api Emails History Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/emails/history/{email_id}": {
      "delete": {
        "tags": [
          "Emails"
        ],
        "summary": "Delete History Entry",
        "description": "Remove one analysed email permanently, not to the trash.",
        "operationId": "delete_history_entry_api_emails_history__email_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails/style-profile": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "Get Style Profile",
        "description": "What the app has learned about how this account writes.\n\nReads the same store the reply generator reads, so what the UI shows and\nwhat the draft actually imitates cannot disagree. It used to call a\nget_storage() that does not exist and swallow the ImportError, so this\nendpoint always answered \"nothing learned\" even with a full corpus.\n\nAn empty corpus is reported as ready=False, not as an invented style.",
        "operationId": "get_style_profile_api_emails_style_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleProfile"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/emails/replies/record": {
      "post": {
        "tags": [
          "Emails"
        ],
        "summary": "Record Reply",
        "description": "Record a reply the user actually sent, so style can learn from it.\n\nOnly delivered mail counts. A draft the user never sent is not evidence of\nhow they write.",
        "operationId": "record_reply_api_emails_replies_record_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/emails/counts": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "Get Email Counts",
        "description": "Return aggregated mailbox counts for sidebar badges and quick stats.",
        "operationId": "get_email_counts_api_emails_counts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/emails/{email_id}": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "Get Email",
        "description": "Fetch details of a specific email.",
        "operationId": "get_email_api_emails__email_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Emails"
        ],
        "summary": "Delete Email",
        "description": "Move email to trash or permanently remove if already in trash.",
        "operationId": "delete_email_api_emails__email_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails/sync": {
      "post": {
        "tags": [
          "Emails"
        ],
        "summary": "Sync Emails",
        "description": "Trigger email sync pipeline and AI categorization.",
        "operationId": "sync_emails_api_emails_sync_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/emails/sync-imap": {
      "post": {
        "tags": [
          "Emails"
        ],
        "summary": "Sync Imap Emails",
        "description": "Directly fetch real emails from Gmail IMAP using App Password.",
        "operationId": "sync_imap_emails_api_emails_sync_imap_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "app_password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails/{email_id}/toggle-read": {
      "post": {
        "tags": [
          "Emails"
        ],
        "summary": "Toggle Read",
        "description": "Toggle read/unread status.",
        "operationId": "toggle_read_api_emails__email_id__toggle_read_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyzeRequest": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject",
            "default": ""
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "sender_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Name"
          },
          "sender_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Email"
          },
          "save": {
            "type": "boolean",
            "title": "Save",
            "default": true
          },
          "generate_reply": {
            "type": "boolean",
            "title": "Generate Reply",
            "default": false
          },
          "tone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tone"
          },
          "personalize": {
            "type": "boolean",
            "title": "Personalize",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "AnalyzeRequest",
        "description": "An arbitrary email supplied by the user, not necessarily from a mailbox."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}